Add HTTPMethod into a separate file
