+++ Gymondo/Gymondo/Networking/EndpointRouter.swift
@@ -55,10 +55,3 @@ extension EndpointRouter {
         return request
     }
 }
-
-enum HTTPMethod: String {
-    case get     = "GET"
-    case post    = "POST"
-    case put     = "PUT"
-    case delete  = "DELETE"
-}
