+++ Gymondo/Gymondo/Networking/HTTPMethod.swift
@@ -1,0 +1,15 @@
+//
+//  HTTPMethod.swift
+//  Gymondo
+//
+//  Created by Ali Elsokary on 22/08/2023.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get     = "GET"
+    case post    = "POST"
+    case put     = "PUT"
+    case delete  = "DELETE"
+}
